// src/app/models/_core/state.ts
/**
 * ID: bh-us-state
 * Name: BH US State
 * Description: State model with stateCode and stateName; used with HelperService's state list
 * Version: 1
 *
 * ==============================
 * Change Log
 * ==============================
 * 2021-07-02 - MW - v1: Initial dev
 */
export interface State {
    stateCode?: string;
    stateName?: string;
}